Ensure that the session (UI) locale stays the same even if receipt locale is different.
This partially fixes an odd bug where a person might use a public contribution page in
French. CiviCRM will create a new contact record and might set its preferred language
to English. The receipt gets sent in English, and then the user is sent to a confirmation
page in English (hence switching the UI locale).

This is only a partial fix, because if the user was on a French contribution page,
we would expect the tax receipt to be also in French, no matter which language is on
their contact record.